clean home by moving directionRequest to serchButton
