+++ src/components/search/searchButton.js
@@ -1,11 +1,14 @@
-import React from "react";
+import React, { useState } from "react";
 import {
     getGeocode,
     getLatLng,
 } from "use-places-autocomplete";
+import { DirectionsService } from '@react-google-maps/api';
 import axios from "axios";
 
+
 const doDirectionRequest = (startLocation, destination, setRoute) => {
+
     const requestBody = {
         startLocation: startLocation,
         destination: destination,
@@ -24,25 +27,53 @@ const doDirectionRequest = (startLocation, destination, setRoute) => {
         });
 }
 
-function SearchButton({ to, from, setRoute }) {
+function SearchButton({ to, from, routeResponse, setRouteResponse }) {
+
+    const [route, setRoute] = useState();
+
+    const DirectionsServiceOption = route && {
+        destination: route.destination,
+        origin: route.startLocation,
+        travelMode: "WALKING",
+    };
+
+    const directionsCallback = (response) => {
+        console.log(response);
+
+        if (response !== null) {
+            if (response.status === "OK") {
+                console.log("response: ", response);
+                setRouteResponse(response)
+            } else {
+                console.log("response: ", response);
+            }
+        }
+    };
+
     return (
-        <button type="button" class="btn btn-outline-danger"
-            onClick={() => {
-                console.log("from: " + from)
-                console.log("to: " + to)
-                getGeocode({ address: from }).then((response) =>
-                    getLatLng(response[0]).then((startLocation) => {
-                        getGeocode({ address: to }).then((response) =>
-                            getLatLng(response[0]).then((destinationCoordinates) => {
-                                doDirectionRequest(startLocation, destinationCoordinates, setRoute)
-                            })
-                        );
-                    })
-                )
-            }}
-        >
-            Way To Go
-        </button>
+        <div>
+            {DirectionsServiceOption && !routeResponse && <DirectionsService
+                options={DirectionsServiceOption}
+                callback={directionsCallback}
+            />}
+            <button type="button" class="btn btn-outline-danger"
+                onClick={() => {
+                    console.log("from: " + from)
+                    console.log("to: " + to)
+                    getGeocode({ address: from }).then((response) =>
+                        getLatLng(response[0]).then((startLocation) => {
+                            getGeocode({ address: to }).then((response) =>
+                                getLatLng(response[0]).then((destinationCoordinates) => {
+                                    doDirectionRequest(startLocation, destinationCoordinates, setRoute)
+                                })
+                            );
+                        })
+                    )
+                }}
+            >
+                Way To Go
+            </button>
+        </div>
     )
 }
