+++ src/pages/Home.js
@@ -5,7 +5,6 @@ import Search from "../components/search/searchInput";
 import Map from "../components/map/map"
 import RangeSlider from "../components/search/timeRangeSlider";
 import SearchButton from "../components/search/searchButton";
-import { DirectionsService } from '@react-google-maps/api';
 import "./Home.css";
 
 const libraries = ["places"];
@@ -30,27 +29,7 @@ function Home() {
   const [to, setTo] = useState("");
   const [from, setFrom] = useState("");
   const [time, setTime] = useState("");
-  const [route, setRoute] = useState();
   const [routeResponse, setRouteResponse] = useState();
-
-  const directionsCallback = (response) => {
-    console.log(response);
-
-    if (response !== null) {
-      if (response.status === "OK") {
-        console.log("response: ", response);
-        setRouteResponse(response)
-      } else {
-        console.log("response: ", response);
-      }
-    }
-  };
-
-  const DirectionsServiceOption = route && {
-    destination: route.destination,
-    origin: route.startLocation,
-    travelMode: "WALKING",
-  };
 
   return (
     <div>
@@ -80,17 +59,14 @@ function Home() {
           </div>
 
           <div class="searchDirectionButton">
-            {isLoaded && <SearchButton to={to} from={from} setRoute={setRoute} />}
+            {isLoaded && <SearchButton to={to} from={from} routeResponse={routeResponse} setRouteResponse={setRouteResponse} />}
           </div>
 
         </form>
       </div>
       {isLoaded && <div>
         <Map onMapLoad={onMapLoad} routeResponse={routeResponse}></Map>
-        {DirectionsServiceOption && !routeResponse && <DirectionsService
-          options={DirectionsServiceOption}
-          callback={directionsCallback}
-        />} </div>}
+      </div>}
     </div>
   );
 }
